LayoutTests/inspector/css/getMatchedStylesForNodeContainerGrouping-expected.txt

Tests for the CSS.getMatchedStyleForNode command and container rule groups.


== Running test suite: CSS.getMatchedStyleForNode.ContainerGrouping
-- Running test case: CSS.getMatchedStyleForNode.ContainerGrouping.Narrow
PASS: Should have 1 authored rules.
- Testing rule #0
PASS: Selector text should be "#item".
PASS: "color" property value should be "peachpuff".
PASS: Source code for rule should be in file named "getMatchedStylesForNodeContainerGrouping.html".
PASS: Rule should have no groupings.

-- Running test case: CSS.getMatchedStyleForNode.ContainerGrouping.Medium
PASS: Should have 2 authored rules.
- Testing rule #0
PASS: Selector text should be "#item".
PASS: "color" property value should be "lemonchiffon".
PASS: Source code for rule should be in file named "getMatchedStylesForNodeContainerGrouping.html".
PASS: Rule should have 2 grouping(s).
PASS: Grouping 0 should have a type of "container-rule".
PASS: Grouping 0 should have a text of "usedName (width >= 200px)".
PASS: Grouping 1 should have a type of "supports-rule".
PASS: Grouping 1 should have a text of "(color: red)".
- Testing rule #1
PASS: Selector text should be "#item".
PASS: "color" property value should be "peachpuff".
PASS: Source code for rule should be in file named "getMatchedStylesForNodeContainerGrouping.html".
PASS: Rule should have no groupings.

-- Running test case: CSS.getMatchedStyleForNode.ContainerGrouping.Wide
PASS: Should have 3 authored rules.
- Testing rule #0
PASS: Selector text should be "#item".
PASS: "color" property value should be "thistle".
PASS: Source code for rule should be in file named "getMatchedStylesForNodeContainerGrouping.html".
PASS: Rule should have 1 grouping(s).
PASS: Grouping 0 should have a type of "container-rule".
PASS: Grouping 0 should have a text of "(width >= 300px)".
- Testing rule #1
PASS: Selector text should be "#item".
PASS: "color" property value should be "lemonchiffon".
PASS: Source code for rule should be in file named "getMatchedStylesForNodeContainerGrouping.html".
PASS: Rule should have 2 grouping(s).
PASS: Grouping 0 should have a type of "container-rule".
PASS: Grouping 0 should have a text of "usedName (width >= 200px)".
PASS: Grouping 1 should have a type of "supports-rule".
PASS: Grouping 1 should have a text of "(color: red)".
- Testing rule #2
PASS: Selector text should be "#item".
PASS: "color" property value should be "peachpuff".
PASS: Source code for rule should be in file named "getMatchedStylesForNodeContainerGrouping.html".
PASS: Rule should have no groupings.
